feat: made window aware of its project
In a project meeting it was decided that each opened project should have
their own AuditWindow. It is now aware of the projects base directory,
and only adds violations that start with the same path.

Added formatting for a more explicit title if multiple windows are open.

Also implemented a base structure for SplitPane idea, where one pane
contains the violations/overview. The other pane is responsible for
showing the ruledescriptions if there are any.