feat: implemented a PackageEventHandler
Implemented a PackageEventHandler that should handle Package Events
emitted by the BlueJ runtime. It's responsibilities are spawning/opening
windows when a new project is opened. It also closes the window manually
if the User has not closed it themselves when a package is closed.

This replaces the functionality previously provided in
AppEventHandler.java