feat: add checking of all class files in the opened project
Also added a guard on the method to prevent the non compiled classes
from being run checks on. Also explicitly commented out exception
handling because they should in theory never happen because a BlueJ
project will presumably always be open when this class calls methods
requiring a Project/Package to be found/open.