feat: allow adding multiple config files
Moved from a single custom config file to allowing multiple config files
to be added. They are stored as key-value pairs of config name and the
config path. This commit implements all the necessary features for it to
function. Also implemented methods for setting the current
configuration, which will not override the selected default
configuration but only be active during the current runtime.